Rename translator/generator files
